+++ er/microcontrole.sql
@@ -14,14 +14,14 @@ INSERT INTO microcontrole.tipo_operacao (code, nome) VALUES
 
 CREATE TABLE microcontrole.monitoramento_feicao(
   id SERIAL NOT NULL PRIMARY KEY,
-  tipo_operacao_id INTEGER NOT NULL REFERENCES microcontrole.tipo_operacao (code),
-  camada_id INTEGER NOT NULL REFERENCES macrocontrole.camada (id),
+  tipo_operacao_id SMALLINT NOT NULL REFERENCES microcontrole.tipo_operacao (code),
+  camada_id SMALLINT NOT NULL REFERENCES macrocontrole.camada (id),
   quantidade integer NOT NULL,
   comprimento real,
   vertices integer,
   data timestamp with time zone NOT NULL,
-  atividade_id INTEGER NOT NULL REFERENCES macrocontrole.atividade (id),
-  usuario_id INTEGER NOT NULL REFERENCES dgeo.usuario (id)
+  atividade_id SMALLINT NOT NULL REFERENCES macrocontrole.atividade (id),
+  usuario_id SMALLINT NOT NULL REFERENCES dgeo.usuario (id)
 );
 
 CREATE INDEX monitoramento_feicao_idx
@@ -34,8 +34,8 @@ CREATE TABLE microcontrole.monitoramento_apontamento(
   quantidade integer NOT NULL,
   categoria VARCHAR(255) NOT NULL,
   data timestamp with time zone NOT NULL,
-  atividade_id INTEGER NOT NULL REFERENCES macrocontrole.atividade (id),
-  usuario_id INTEGER NOT NULL REFERENCES dgeo.usuario (id)
+  atividade_id SMALLINT NOT NULL REFERENCES macrocontrole.atividade (id),
+  usuario_id SMALLINT NOT NULL REFERENCES dgeo.usuario (id)
 );
 
 CREATE INDEX monitoramento_apontamento_idx
@@ -47,8 +47,8 @@ CREATE TABLE microcontrole.monitoramento_tela(
   id SERIAL NOT NULL PRIMARY KEY,
   data timestamp with time zone NOT NULL,
   zoom REAL NOT NULL,
-  atividade_id INTEGER NOT NULL REFERENCES macrocontrole.atividade (id),
-  usuario_id INTEGER NOT NULL REFERENCES dgeo.usuario (id),
+  atividade_id SMALLINT NOT NULL REFERENCES macrocontrole.atividade (id),
+  usuario_id SMALLINT NOT NULL REFERENCES dgeo.usuario (id),
   geom geometry(POLYGON, 4326) NOT NULL
 );
 
@@ -68,8 +68,8 @@ CREATE INDEX monitoramento_comportamento_atividade_id_idx ON microcontrole.monit
 CREATE TABLE microcontrole.monitoramento_comportamento(
   id SERIAL NOT NULL PRIMARY KEY,
   data timestamp with time zone NOT NULL,
-  atividade_id INTEGER NOT NULL REFERENCES macrocontrole.atividade (id),
-  usuario_id INTEGER NOT NULL REFERENCES dgeo.usuario (id),
+  atividade_id SMALLINT NOT NULL REFERENCES macrocontrole.atividade (id),
+  usuario_id SMALLINT NOT NULL REFERENCES dgeo.usuario (id),
   propriedade VARCHAR(255) NOT NULL,
   valor VARCHAR(255) NOT NULL
 );
